refactor(promql): fold the shared engine test fixtures into mod.rs
The sibling files' tests import them from crate::engine::tests instead
of a dedicated test_util file.
